refactor: extract get_valid_properties into shared module
